requirements.txt
fastapi<1
psycopg2-binary<3
uvicorn<1
